pyproject.toml
[project]
name = "masterthesis-data-generation"
version = "0.1.0"
description = "Kubric/PyBullet pipeline for generating synthetic friction physics video datasets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "pillow>=12.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
generate-clip = "experiments.scripts.generate_single_clip:main"
generate-dataset = "experiments.scripts.manager_benchmark:main"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]
